cli: change convert_klut to strash
