+++ cli/mockturtle_exe.cpp
@@ -53,7 +53,7 @@
 
 /* include mapping commands */
 #include "commands/mapping/emap.hpp"
-#include "commands/mapping/convert_klut.hpp"
+#include "commands/mapping/strash.hpp"
 
 /* include printing commands */
 #include "commands/printing/print_stats.hpp"
